+++ Models/Study/Specialty.cs
@@ -4,10 +4,10 @@ namespace Dotnet.Models
 {
 	public class Specialty
 	{
-		public int Id { get; set; }
+		public ulong Id { get; set; }
 		public string Name { get; set; }
 		public string Code { get; set; }
-		public int FacultyId { get; set; }
+		public ulong FacultyId { get; set; }
 		public Faculty Faculty { get; set; }
 		public List<StudyGroup> StudyGroups { get; set; } = new List<StudyGroup>();
 	}
