+++ Models/Study/Teacher.cs
@@ -5,8 +5,8 @@ namespace Dotnet.Models
 {
 	public class Teacher
 	{
-		public int Id { get; set; }
-		public int UserId { get; set; }	
+		public ulong Id { get; set; }
+		public ulong UserId { get; set; }	
 		private string specialization;
 		public string Specialization 
 		{
