+++ ViewModels/API/Education/SpecialtyViewModel.cs
@@ -2,7 +2,7 @@ namespace Dotnet.ViewModels.API.Education
 {
 	public class SpecialtyViewModel
 	{
-		public int Id { get; set; } = 0;
+		public ulong Id { get; set; } = 0;
 
 		public string Name { get; set; } = "";
