+++ ViewModels/API/Education/StudentViewModel.cs
@@ -2,9 +2,9 @@ namespace Dotnet.ViewModels.API.Education
 {
 	public class StudentViewModel
 	{
-        public int Id { get; set; } = 0;
+        public ulong Id { get; set; } = 0;
 
-        public int UserId { get; set; } = 0;
+        public ulong UserId { get; set; } = 0;
 
         public bool IsLearns { get; set; } = true;
