+++ ViewModels/WebApp/Student/StudentViewModel.cs
@@ -5,13 +5,13 @@ namespace Dotnet.ViewModels.WebApp.Student
 	public class StudentViewModel
 	{
 		[Required(ErrorMessage = "Укажите пользователя.")]
-		public int UserId { get; set;}	
+		public ulong UserId { get; set;}	
 		
 		[Required(ErrorMessage = "Укажите статус.")]
 		public bool IsLearns { get; set; }
 
 		[Required(ErrorMessage = "Укажите учебную подгруппу.")]
-		public int StudySubgroupId { get; set; }
+		public ulong StudySubgroupId { get; set; }
 
 		public bool IsOnline { get; set; } = false;
 	}
